Extract exception handling
