+++ lib/orbf/rules_engine.rb
@@ -13,6 +13,7 @@ require_relative "./rules_engine/log"
 require_relative "./rules_engine/services/tokenizer"
 require_relative "./rules_engine/services/period_converter"
 require_relative "./rules_engine/services/period_iterator"
+require_relative "./rules_engine/services/solver_error_handler"
 require_relative "./rules_engine/services/solver"
 require_relative "./rules_engine/services/value_formatter"
 require_relative "./rules_engine/services/codifier"
