+++ lib/orbf/rules_engine/services/solver.rb
@@ -56,51 +56,18 @@ module Orbf
           split_problem(problem, calc)
           @solution = calc.solve(equations) do |error|
             if error.is_a?(Dentaku::UnboundVariableError)
-              on_unbound_variable_error(error, problem)
+              Orbf::RulesEngine::SolverErrorHandler.on_unbound_variable_error(
+                error,
+                problem,
+                @variables
+              )
             else
-              on_error(error, problem)
+              Orbf::RulesEngine::SolverErrorHandler.on_error(error, problem, @variables)
             end
             raise error
           end
         end
         @solution
-      end
-
-      def on_error(error, problem)
-        puts [
-          ColorizedString["----------- ERROR !!!"].colorize(:red),
-          field_message("  Message            : ", error.message),
-          field_message("  Problem            : ", JSON.generate(problem)),
-          ColorizedString["--------------------"].colorize(:red)
-        ].join("\n")
-      end
-
-      def on_unbound_variable_error(missing_var_error, problem)
-        missing_var = @variables.index_by(&:key)[missing_var_error.recipient_variable]
-        puts [
-          ColorizedString["----------- ERROR !!!"].colorize(:red),
-          field_message("  Message            : ", missing_var_error.message),
-          field_message("  Recipient_variable : ", missing_var_error.recipient_variable.to_s),
-          field_message("  Unbound_variables  : ", missing_var_error.unbound_variables.to_s),
-          field_message("  Variables          : ", "\n" + missing_var.to_s),
-          field_message("  Equation           : ", "\n#{highlight(problem[missing_var_error.recipient_variable], missing_var_error.unbound_variables)}"),
-          field_message("  Formula expression : ", "\n" + missing_var&.formula&.expression.to_s),
-          ColorizedString["--------------------"].colorize(:red)
-        ].join("\n")
-      end
-
-      def field_message(field_name, message)
-        [
-          ColorizedString[field_name].colorize(:light_blue),
-          message
-        ].join(" ")
-      end
-
-      def highlight(equation, unbound_variables)
-        tokens = Tokenizer.tokenize(equation).map do |token|
-          unbound_variables.include?(token) ? ColorizedString[token].colorize(:yellow) : token
-        end
-        tokens.join("")
       end
 
       def split_problem(problem, calc)
@@ -113,6 +80,7 @@ module Orbf
         end
       end
 
+      # rubocop:disable Rails/TimeZone
       def benchmark(message)
         start = Time.now
         value = nil
